+++ Math/Math/GPUSparseMatrix.cu
@@ -463,7 +463,7 @@ namespace Microsoft { namespace MSR { namespace CNTK {
     }
 
     template<class ElemType>
-    void GPUSparseMatrix<ElemType>::Resize(const size_t numRows, const size_t numCols, int size)
+    void GPUSparseMatrix<ElemType>::Resize(const size_t numRows, const size_t numCols, size_t size)
     {               
         m_nz = 0; 
         m_colIdx = -1;
